+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### Compatible changes
 
+- when given a scope, do not load all records to memory during validation
 
 ## 0.16.1 - 2019-05-14